added config package for data persistence
